hands_on_activities/Group 1/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
pyttsx3>=2.90
pyaudio>=0.2.11
requests>=2.28.0
wave
typing
